Use Gag crate
Was double running ecah day's code, once with output for the stars,
and once without so I could run it multiple times and benchmark the
runtime without spamming the output. Needless to say that was a
hopless duplication and the Gag crate allows me to run the main function
whilst suppressing Stdout. Much happier now