2022 Day 15 reworked
OK now I'm happy with this. Rather than a scan search we use the fact
that the only free square must be one more than distance ebtween two
pairs of sensors. This cuts the search down drastically

I cheated a bit on the test case as there's actually a few points
returned, but the first is the right answer. To be more thorough we
shoudl filter out multiple points by doing a scan against each sensor
and filtering out any found points that are inside a third sensor. Dont
need to do this on the real data though as there's only two pairs of
points, which then uniquely find the solution