2015, Day 7
Only tricky thing here was memoizing the values of each gate as a
recursive implementation required mulitple mutable references on the
same circuit array. This would have always been fine as we were never
mutating the same element (assuming we don't have a circular circuit),
however Rust can't see that. Could have got round it with unsafe Rust or
RC & Mutex

In the ed though I decided the better strategy was to to sort the
circuit, with each gate only relying on gates to the left of it. This
allowed me to split slices of the cicuit and not run into multiple
mutable refernces.