2015 Day 11
I actually solved the problem in my head whilst readnig the problem
page, it wasn't that hard. For completeness I have written code that
will do the same. This is not perfect as it won't deal with the final
example input correctly, although it wouldn't be that much work to fix