2015 Day 12
I have a strong suspicion we were intended to read the string as a JSON
object and then owrk with that. Beacuse the only library I'm using is
itertools, that wasn't an option so had to do some funky searching
fowrdas and backwards through strings and arrays of bytes.

Couldn't think of a neat way round an allocated string as the
intermediate result from having stripped any object containing "red".
Still the program ran fast enough