2015 Day 19 Part 1 only
It doesn't work for the example input (as explained) in the
comments for test one.

I have made life much harder for myself by not allocating and just
analysing the strings in situ in the binary code. Having now seen part
2, where unsuprpsingly we need to do a series of substitutions this
method won't help us at all and I'll need to do the "natural" solution
by creating a series of allocated strings for each iteration of
transformations