2015 Day 19 Part 2
This was a beast! The naive BFS approach rather predictably blew up so I
had to look at what the inputs were actually doing. It turns out that
each transform input was one of a few subsets:

e => _._
_ => _._
_ => _.Rn._.Ar
_ => _.Rn._.Y._.Ar
_ => C.Rn._.Ar
_ => C.Rn._.Y._.Ar
_ => C.Rn._.Y._.Y._.Ar

In the examples above, _ is used to indicate any of the base elements:
Al, B, Ca, F, H, Mg, N, O, P, Si, Th, Ti
The dots . are used for clarity but are not in the substitutions

It can therefore be seen that each transformations adds (or removes in
reverse) 0, 1 or 2 base elements to the transformation molecule. It is
easier to reason about the manufacture of the molecule in reverse i.e.
how many steps it takes to remove all but one element.

The number of base elements removed with each transformation depends on
which pattern is used and note that the patterns with Y in them (on the
RHS) increase the base elements removed in one transformation. Therefore
we can conclude that we can remove 1 + count(Y) base elements in each
turn.

However for patterns that start C.Rn, one less base element can be
removed per transformation than the above rule. I only had one instance
of this, right at the start of my input. To be safe, I've coded assuming one
might find a C.Rn pattern anywhere in the input since it is easy enough
to handle.

With all of that we can make a single pass through the input molecule,
counting the base elements, deducting one for each Y found and adding
one for each C.Rn found. I actually coded in a recursive fashion but the
essence remains