Utility to download input files
Because we're not meant to store the inputs in a public repo (see link
below), this repo has no inputs which is pretty annoying on a clean
pull. Have added a little utility sub-crate, in the inputs folder, that
will download all missing input files to the right places.

This utility relies on *your* session ID being stored in an environment
variable called AOC_SESSION. Without this, there's no way to download
your specific inputs (& I'm not giving you mine!)

The code is pretty fragile and doesn't check the status code for a 200
before proceeding. This is where bad session inputs will likely surface so
I may well have to harden this in the future.

https://www.reddit.com/r/adventofcode/comments/zh2hk0/2022friendly_reminder_dont_commit_your_input/