2015 Day 1
Just going back and filling in the first few days of 2015 that I must
have done in Python last year.

This is a simple one, although I couldn't think of a simple way to make
part 2 a pure iterator solution. I did toy with a fold or find but they
both seemed to be getting more involved than the problem merited