2015 Day 4
Bit of a bruteforce jobby. On current stats (only missing day 5), this
one day is taking half the runtime of the entire year!

Could look into the internals of how md5 is produced, remove the
dependency and hopefully shortcut the calculation. Sounds like a lot of
effort though and I'm not particularly motivated to do it