2016 Day 3
Nothing that special here. Relied on iterators and chunks to get the job
done.

Central insight for part 2 was that I could split on whitespace (which
captures both spaces and new lines) and then chunk into groups of 9,
which captures a set of three vertical "triangles". Then use windows to
extract each set of instructions, skipping over the irrelevant
instructions for the other two sets.