2016 Day 5
MD5 hashes were brutal in this code challenge. I couldn't think of a
decent way to short-circuit computing all of them in sequence, so
that's what I did. The one optimisation I used was to pre-compute the
valid hashes into an array that could be re-used for parts one and two

Still takes over 5s to crunch these bad boys though