2016 Day 11
Had a lot of "fun" with this one!

My basic idea was to use bitmasks to represent each floor and then ran
the A* algorthim on the collections of floors to prefer moves that got us
closer to the required end state.

A massive optimisation was to summarise each floor and track the summary
stats for visited state. This works because we don't actually care which
element pair is on each floor and they can be considered interchangable