2016, Day 12
I took some liberties with this one and interpreted what the underlying
assembley was doing, rather than building the assembley interpreter as
question was implying we should.

Turns out it was a Fibonacci expension, plus two other numbers that were
just multiplied together. The assembly was going to do all of that using
increment counts which would have been mighty slow. Programming it as
descibed at the higher level, is obviously much faster.

I've taken a bit of a flier and assumed that the only way different
inputs will vary is which Fibonacci number you need to go to (which
varies for the two parts) and the two other factors that are multiplied
and added in at the end. If this assumption is not true then the code
won't be universally applicable but I think it quite likely it wil be