2016 Day 17
A fairly straight forward day. Used an A* search for part 1, with no
need to track the visited states since each entry has a unique path
history, which means a different set of doors will be available to it.

The second part just required us to exhaust the pool of all possible
paths and record the longest one. This was a small tweak on part 1. I
even kept the A*, which likely slowed me down a bit, as every path needs
to be searched anyway, so the order of looking is not that important