2016 Day 18
Found this day a bit simpler than the previous few.

I represented each row as a u128 (the input was 100 tiles wide). I used
a 1 for a trap and 0 for safe.
To generate a successive row the centre tile doesn't matter, and
the requirements boil down to a trap being present if there's a trap in
one space to either side and a safe tile on the other side.
We can use bit sifting to calculate this quickly and efficiently:

1) Shift 1 space left for the right space values
2) Shift one space right for the left space values
2) XOR (^) the above to test for one of them being 1 and the other 0

This operation has the nice side effect of being done to the whole row
in one step.

I needed to keep around a mask to make sure values to the left of the
actual data were kept to zero but other than that there wasn't much to
it