2016 Day 19
The solution looks nothing like the problem statement! That's because I
analysed how the two parts progressed as more elves were added.

Part 1 was equal to 1 for any number of elves that fell on a power of 2
(1, 2, 4, 8, 16, ...). Between these resets the resultant elf stepped up
in 2's. For example:
  P1(4) = 1 <- power of 2
  P1(5) = 3
  P1(6) = 5
  P1(7) = 7
  P1(8) = 1 <- power of 2
  P1(9) = 3
  ...
So to calculate part 1 we just need to determine which power of 2
segment we're on and the result is 1 + 2 * however far into the segment
we are

Part 2 also had a cyclical pattern. Here the pattern repeated over powers
of 3. The part 2 was equal to the number of elves (i.e. the last elf)
for any power of 3 (1, 3, 9, 27, 81, ...). The value reset to 1 for one
greater than any power of 3. Between these resets, the function
incremented in 1's for the first half of the interval (which also
happens to be the previous power of 3 lower bound) and then increments
in 2's for the second half. For Example:
  P2(3)  = 3 <- power of 3
  P2(4)  = 1 <- reset to 1
  P2(5)  = 2
  P2(6)  = 3
  P2(7)  = 5 <- second half of interval, now increment in 2's
  P2(8)  = 7
  P2(9)  = 9 <- power of 3
  P2(10) = 1 <- reset to 1
  ...

You can prove both these series by induction. With formulae in hand the
code became VERY simple :)