2016 Day 21
Not too tricky a day. I went with a heavily type-stated solution,
parsing things from strings and proper validations on the bounds.

Part 2 was mostly trivial. The only part requirin any thought was
instruction to reverse based on the index of a letter. By inspection, it
can be seen the the forward operation had a unique 1 to 1 mapping, so
the reverse just required us to invert that index mapping. A lot of th
other operations remained identical (swapping, for example) and the
rotates just required us to invert the direction. One last step I
stupidly forgot was we need to iterate through the instructions
themselves in reverse, but easy enough to fix once identified!