2016 Day 22
This problem was definitely designed to be solved by hand rather than by
computer.

The first part was easy enough, and even easier once the part 2
constraints were revealed. I limited the possible from and to nodes by
looking at the max available and min used and filtering out nodes that
couldn't possibly be on one side or another. Then it was a simple
Cartesian product (thanks Itertools!) to check whether the pair actually
met the criteria

The second part could be solved quite simply by inspection, if you print
out the filesystem. However in the spirit of coding up a solution, I
moved the empty node up the one node to the left of the target node
using a BFS search. Then I applied repeated sets of 5 moves to shift the
target over to the left, using as many as the filesystem is wide.