2016 Day 24
There's a bit to unpack here. My high-level plan was first to build up a
weights matrix of the minimum distance between each pair of nodes. Then
with the 8x8 weights matrix this becomes a Travelling Salesman Problem.

The matrix was built up using an undirected BFS from each start point. I
didn't direct as we wanted to find all destinations from each start
point.

An optimisation I employed on the initial BFS search was to treat the
input data a long string of bytes, with a fixed column width. To find
the thing at a given cell we can just look (row * width) + column into
the byte array (remembering to account for the newline byte)

Another optimisation was to run the searches from in parallel for each
start node.

Then with the weights matrix I could have just iterated through the 7!
and 8! permutations to find the minimum but I built a branch and prune
algorithm for a bit more efficiency

All told it runs in about 700 us on my machine. Noice!