2017 Day 2
Not much of interest to report. Another day with iterators

Only mildly interesting thing was for part 2 where I sorted the iterator
of numbers (as big numbers will always only ever be divided by smaller
numbers). I also needed to clone the sorted iterator for the inner loop
so that the position of the first item search was preserved, otherwise
the find call would consume the out iterator