2017 Day 3
Part 1 - I worked out which ring we'd be on first, as each of those
would require a step in, followed by how many steps away from the middle
of the side we are. I used the insight that the final value of each ring
was stepping up in the series:
  (2n - 1)^2 i.e. 1, 3*3 = 9, 5*5 = 25, 7*7 = 49
The answer was just the sum of those two parts

Part 2 - There was no obvious series that I could see. Helpfully the
series is expanding exponentially, so we didn't need to go through many
steps of the problem as described to get to the answer. I therefore
modelled the grid of values as a 2D array of numbers and just iterated
through the problem until the current value was bigger than the target