2017 Day 4
This has had a few iterations for performance. In the end, what worked
best was a couple of nested array structures stored on the stack:

 - SplitPassword is an array of Words, with a tracker of how much of the
   array is actually being used (to make it like a Vec). I went a bit
   fancy and implemented Iterator for it to make looping through more
   ergonomic
 - Word is an array of u8 or bytes. This i just kept as is as the
   trailing zeroes don't affect the logic

For part 2, we just need to make sure that the Words that are stored are
sorted first. Then we can run an equivalence test on the Words to find
out if they are anagrams of one another

A small perf tweak was to make sure the words were sorted in reverse
order. This means the first digits of the word weren't the trailing
zeroes and so we search the significant bytes of the words first

The repeated_parts function does a simple triangle search, popping off
the end item and then comparing that to all other words in the remaining
list. It quits out as soon as we find a match