2017 Day 6
My central idea here was that the configuration of the blocks could be
stored in a single u128. When you split this into 16 parts you have 8
bits for each bank, which can hold a value up to 255. This is more than
enough to specify the state of each bank in any one iteration.

Then it was just case of doing a spot of bit twiddling to read out each
bank, find the max and run a reallocation pass.

For the problem itself, I just stored the u128s in a HashSet (part 1)
and HashMap (part 2), to easily find when we had a repeated value. I
could have halved the runtime by just using a HashMap and spitting out
both parts in a single pass but I didn't fee like it.