2017 Day 7
I seriously over-engineered today's solution. In the end what worked was
making a custom struct for an individual program, that was sortable
based on name. This struct has a iterable collection of children
programs, which are referenced by name (&str).

I stored all these programs in an array and sorted the array (by program
name). This meant I could look up a given program with a binary_search.
The other two data structure candidates would have been:
  * a hash table - but we're not doing that many lookups, so I figured
    the overhead of setting one up wouldn't be worth it
  * a tree - this would require sorting the input data first (which
    we're already doing) plus dealing with trees in Rust isn't much fun

For part 1 I created a candidates vector (I figured one heap allocation
was fine), to this I looped through the programs and added any program
that had children. A program can only be the root if it has children
itself. In the first loop I was also removing from the candidates list
any program that was a child of a later program, as the root program
cannot be a child of any program. I then did a second loop, as a single
pass loop would not correctly exclude any child programs that come after
their parent has been mentioned. After the double loop we're left with
only one data item (thanks well-formed AoC puzzle) so that's our answer.

For part 2, starting with the root program, I recursively found the
total weight of each program plus all their children. Processing the
entire tree would have been wasteful, so I allowed the recursive program
to return a result, with the Err state indicating that the mismatched
node had been found.