2017 Day 8
Fairly simple day. I used a HashMap to store the values of non-zero
registers. To save doing it twice over for the two parts I generated the
HashMap of registers once and recorded the max value in the single pass
loop, for part 2. It made the parts 1 and 2 trivially simple since all
the work was done in the generate_registers function