2017 Day 9
Not much to this one, just read the characters of the input string one
by one and process according to the rules. Had a sperate sub method for
inside garbage as it's a bit different.

I did two passes through the input string, one for each part. This could
have definitely been cut down into a single pass in a pre-processing
step but it was already a quick day and I CBA