2017 Day 10
I tried to be cute in part 1 and only tracked the position of the two
places we care about. I had to run this backwards, seeing what position
0 and 1 ended up in when run backwards. This worked fine and was mighty
fast, but it wasn't applicable to part 2.

For part 2 I decided I had to play it as explained. There's not clever
about my solution to explain really