2017 Day 12
A couple of things worth talking about in today's challenge.

I converted the input into a Vector of &str references. One allocation
isn't that bad. I dropped the leading number and the <-> indicator. The
leading number since the pipes were referenced in order, so the vec
index would keep track of that. I didn't bother decoding the to pipes
since that would either require nested vectors, which would be bad, or
would have required using fixed arrays with wasted space. In my
algorithm I only decode the to string once so storing the str reference
isn't that bad.

The other point of note is that I wrote my own Hash to keep track of the
found pipes. Since the input had pipes from 0 to 1,999, I could keep
track of the found state by using the first 100 bits of 20 u128s. This
was marginally quicker than just using a standard HashSet, but a win's a
win