2017 Day 13
This one was a bit more involved & the code may not be the most obvious
to read

Part 1 was trivial. There were a couple of things in this that equally
apply to part 2:
 * Rather than bother with stepping the system state on as the program
   crosses the firewall, we can instead move each scanner that is `depth`
   away from the start `depth` positions on. This reduces the problem to
   a static state that can be inspected without doing any stepping
 * If the scanners move back and forwards then they hit the zero
   position every `2 * (range - 1)` goes. Since we're only interested in
   when scanners are in the zero position we can turn this problem into
   one modulo `2 * (range -1)`. In the rest of this write up I refer to
   this distance as the cycle of the scanner

Part 2 I initially wrote up where I incremented in steps of 1 and saw if
any scanners blocked us. Of course, the problem was specified such that
we need to do over 3 million iterations to find the first instance where
we're not blocked and this took an unacceptably long time to solve like
this. So another approach was required...

I initially discounted the Chinese Remainder Theorem as an approach for
this as we're looking for the iteration when each scanner is not at
zero. CRT finds the point at which things are equal to zero. However
looking more closely at the input, I realised I had been too hasty.

There was one scanner with cycle (and range) 2. It was in position 1,
meaning it would allow a packet that left instantly and any number of
even wait times, but would block any packet that waited an odd amount of
time. This is relevant as every other scanner started at an even
position, but the presence of this one scanner meant that we only ever
needed to look at the even waits.

The next thing to notice is that there were multiple scanners at
different locations with the same cycle (or range). They were set out
such that for a lot of the cycles, there was a scanner starting every
even location in the cycle _except one position_. If it had covered
every location, then there would be no solution to the problem, since:
 - for every odd wait time we would be blocked by the scanner with cycle
   2,and
 - for every even wait, one of the scanners on that cycle would have
   rotated around into the zero position (remember that every even
   position over the cycle is taken in this hypothetical)

Therefore, for those cycles that had all but one even starting position
covered, the gap starting position is the element that we need to rotate
into the zero position. This is why the CRT does apply, it's just that
the element we need to solve into the zero position is identified by its
absence. This is what my `find_main_cycles` method is doing, it is
finding the cycles that have a singular even gap and then recording what
the collection of all such (cycle, gap) pairs is.

Solving the CRT involves incrementing all of the main cycle pairs,
starting with a step size of 2. When any one pair first aligns the gap
with zero (mod cycle), we can now say that the step size should only
ever go up by a multiple of the cycle that has just been solved, as this
is the only way to keep that gap at zero. So each time we solve a new
gap to zero we:
 - increase the step size: `step *= cycle / 2`. We divide by 2 since
   that factor is in every cycle and it's also the initial step size.
   Note if the cycles we were solving were not co-prime we'd have to do
   a bit more faff here
 - remove the found (cycle, gap) pair from our list
We continue to do this until all gaps are at zero, at which point we
have solved the system for all the main cycles

I did still need to do the traditional solver after this as there were
cycles that weren't fully covered. However with the much larger starting
wait time and the ability to step in increments of over 60k, the problem
was dramatically sped up