2017 Day 14
A lot of this was just re-using day 10's solution to generate the knot
hash. I made some edits to day 10 to make this work more smoothly.

Since we had to compute the same 128 hashes for each part and they were
not an inconsiderable performance hit, I pre-computed all the hashes for
both parts in an initial step.

Looking at part 2, I decided that the best data structure to use would
be a HashSet of tuples of u8, that represent the row and column
respectively of the location of the 1 bits in the resultant hashes.

To calculate the regions, I could 'pop' an item out of the hash and then
recursively remove all of its neighbours, to find each region. Every
time I started again with a new pop represented a new region. Do this
until the HashSet is empty and we're done.