2017 Day 15
Today's problem was simple enough and is really just a brute force test.
I couldn't think of any clever number theory ways of reducing the
problem.

I coded it in s straightforward manner running the generator
incrementing line inside a fold, so that we just kept the current values
in memory at any one time. This ran in a bit over 500 ms, due to the
sheer number of calcs being run.

I didn't like the runtime, so I had a go at parallelising the
calculation. I was only mildly successful with this (and at the expense
of massive increased memory usage) however since it was a little quicker
and I've not really used threading before, I thought it worthwhile to
commit in this code as my final version. If I were for real writing
this, I'm not sure the extra memory usage justifies the slight speed up

My idea was to first generate vectors of all 40 million values of a and
b. Since part 2 was essentially using the same dataset, I could hold
this in memory and use it for parts 1 and 2 alike. I was able to speed
up the generation stage by spinning off the generation of a and b into
separate threads.

Then with the two 40 million data series, calculating the result was a
simple case of zipping them together, and folding down incrementing the
count each time the lower 16 bits were the same. Part 2 used almost
identical logic but I just needed to insert a filter call on each series
before we zipped them so we only took the items that were divisible by 4
and 8 from the respective series.

I did try using channels to send data across threads but 80 million
sends were much slower than just building up the data vector in place