2017 Day 16
A reasonably straightforward day.

I created an enum for the different move types, and then implemented
them on a Promenade struct, which represents the order of the programs
at a given point in time. Rather than store text in the Promonade
struct, I just stored 0 for 'a', 1 for 'b', 2 for 'c' etc. I convert
back to text by using a custom Display implementation.

For part 2, 1 billion was way too many iterations to run straight up.
Instead I used a HashMap to find the first point I get a cycle, meaning
I get back to the same string I've previously seen a number of steps
ago. I called this number of steps a cycle. Because the dance is pure
and is not modified over time, we can skip forward in multiples of cycle
steps and be guaranteed to be in the same position. So we only need skip
forward until we're within a cycle of the target.

We don't even need to do any more dance iterations at this point as the
HashMap we used to find the first cycle also has within it the state of
the programs at every subsequent iteration (up to a full cycle). It's a
little annoying as we want to find the key for a given value, which is
not really how you're meant to use HashMaps, but we can hack it by
iterating through the key value pairs stopping when we find the right
one.