2017 Day 17
I started today's problem with the insight that the nth insert would be
at position p, where `p = (last_p + cycle) % size + 1`.

For part 1 I used this to build an array of the initial insert position
of all 2018 entries. However this is not the final position as once
inserted an entry can be bumped up by a later insert of an element at
the same or lower location. So we then needed to loop through adjusting
all the entries that got subsequently bumped.

For part 2, we only needed to know the position of the entry in the 1
location. Since the zero value always remains in the zero position, this
is equivalent to looping through all 50M steps and finding the last time
that an initial position of 1 is found (all earlier instances will
ultimately get bumped up out of position 1).

We can do this and the calc ran fine, but for 50M iterations it was a
little slow (like half a second). So I looked to batch up the iterations
into groups that would trigger the modulo operation, as all the
intermediate steps could calculated as:
`p = last_p + batch_size * (cycle + 1)`