2017 Day 18
This was a really fun day! The logic of the assembly was such that I
didn't bother trying to analyse it to skip loops as there didn't seem to
be many big cycles in it.

I wrote a straight-up interpreter as directed by the problem and it ran
plenty fast enough for my requirements.

In part 2 we were meant to be running two copies of the program in
parallel. Even though I suspect it was totally possible to run them in
interleaved series, switching over each time one caps out waiting for
new messages, I went all in and ran them in seperate threads with
channels just like the problem suggested we should.

The only slightly tricky thing was detecting the deadlock state. To
break this, I ran a third thread that kept track of the message queue
state in the other two threads and which would therefore know when a
deadlock condition had been reached. This worked fine and so the problem
was done