2017 Day 20
Quite a lot of boilerpalte today to model particles and be able to
compare them for sorting.

For part 1, the cloest is always going to be a particle with the lowest
acceleration, in the long run. In my dataset there were several
particles with the lowest acceleration. Unfortunately you can't pull the
same trick with velocity, to tell these apart as the acceleration may
be working against the initial velocity. To make things simple, I just
went 1,000 steps away and saw which had the lowest velocity after that.

I wasn't the best pleased with the algorithm I used for part 2, just
looping for a fixed 500 steps, but I couldn't think of a simple stopping
test.

I did consider solving the crossing points between each pair of
particles. This would require solving simultaneous quadratic equations
for every pair of particles. I stayed away from this in the end as that
sounded pretty involved and not necessarily a performance win, as
there's no simple integer formula for solving quadratics.