Better functional code
I meant to try to remove the for loop before committing but forgot. This
is a small commit to update to better functional code (at least as far
as I'm concerned). Now I don't have to write a for loop and can instead
stay in the iterator syntax for the whole function call