2017 Day 22
I had a couple of mis-steps with today's problem.

Firstly, I tried the nodes as a 256x256 2D stack array. This worked fine
to start with in the 20 iteration initial example but overflowed even
getting to the 10k iterations in part 1

Next I expanded the array size but this generated a stack overflow

So I moved to using HashSets, as most of the nodes will be and remain
clean & and we don't need to store those anywhere. This worked and
solved both parts but the massive number of allocations needed made it
slow

I then moved to a Vector of Vectors i.e. heap allocating my stack
overflow attempt. This was much more performant at around 300 ms on my
machine. However I still wasn't happy and my final optimisation was to
move to a Vector of arrays. I assume this forces the heap memory to be
contiguous rather than vector of pointers to other vectors with each of
the row info. This got my final run-time down below 200ms

I was wondering if the NodeStatus type was a bit wasteful, since it only
has 4 states. It could be expressed in as little as 2 bits, allowing the
nodes row arrays to be u8, each one representing a run of 4 nodes. But I
haven't coded this up

The rest of the problem was fairly standard modelling out of the problem
as stated without any special trickery