2017 Day 23
This was two very different parts.

Part 1, I just copied across the code from day 18 as the problem
suggested we should, made some minor edits and was done.

Part 2 couldn't run through the interpreter in sensible time. So I
analysed the assembly and realised it was counting the number of
non-prime numbers between a pair of largish numbers, but only looking at
numbers at an increment step. The code for that was trivial and I
extracted all the interesting numbers from the input assembly that set
the two large numbers to search between as well as the step size.