Refactor Day 22
It's not a huge win but I've moved to using u8s, storing 4 states in one
cell, rather than an enum per slot, which is a whole 8 bytes by itself

Runtime has come down a bit but not by much. Most of the runtime is in
my NodeGroup rotate function, a simple (but wrong!) version of that
brings runtime down to 40 ms. I think how to do it better so will leave
this as en exercise for the future reader