2017 Day 25
Most of this problem was parsing the input!

Once I did that the problem was pretty simple. I wanted to use bit
operations but there were going to be more than 128 written numbers in
either direction and dealing with the faff of pushing bits over a vec of
u128 didn't seem worth it

I did use two vecs (of bools) for the memory of what's left and right of
us on the tape. Rather than change position I just pushed and popped
from the right and left vecs with each step which did the same job