2018 Day 5
The essence of this solution was actually an answer on the Reddit
solution megathread. I initially had the idea of using a struct that had
a &[u8] reference to the bytes of the input string and a matching array
of booleans to track if the matching unit was included or not.

There were two key innovations that this solution improved. Using a Vec
was much more straightforward as we could push and pop from the end as
we considered each new unit, plus the last method allows for quick
inspection of what's currently on the end.

The other big win was the observation that the filtering for part 1 was
required work for any of the part 2 checks and so could be done in a
single step first (and then not re-done 26 times).

I have used iterators and folds, because I like them. In testing though
using straight up for loops was a bit quicker (~10%). It's a price I'm
willing to pay