2023 Day 1
Nothing too hard here although I took longer than I should as I felt a
bit rusty (no pun!)

I searched in from each end, treating each line as a byte array. Finding
the written numbers was tricky with this and I resorted to using a hard
coded array of the bytes that spell out each of the digits. To match
them I resorted to indexing into the line byte array, finding a slice of
matching length and then checking if that was equal to each digit byte
array representation