2023 Day 3
I treated the input string as one long array of bytes. I iterated
through the string in a single pass. The purpose of this pass was to
firstly identify the numbers. I did this with an optional index that
marked when the current number was first found, or None if we're not in a
number. When I detected the end of the number, which is by reading a
non-number byte, I then searched around the number for any symbols

For part 2, I used the same pass. If we find a number, I searched the
whole way round the number for all instances of an asterisk. I built a
shadow array with the same dimensions as the input. At each location,
which corresponds to the potential locations of the gear asterisks, I
held a custom enum. The enum held which & how many parts had been
observed next to a gear in that location. For the final answer it was a
simple matter of reducing the array down to the two parters and doing
the sum product