2023 Day 4
Another reasonably straightforward day. I parsed each line into how many
winning numbers that scratchcard had. Since this was the time-consuming
bit of the problem, I did this as an initial parsing step, collecting
the wins into a vector.

For part 2 I built a custom stack (which really should have been called
a queue now I think about it). Since the most we could be looking
forwards is 10 cards (the cards can win at most 10 times and we only
increment up the card count over the range of the last win number), I
built the stack with 12 items.
* The first represents the number of copies of the current scratchcard
  we have
* The next 10 are the card counts for the next 10 scratchcards we are
  due to draw
* The final slot is to feed the incoming cards with the original card
  count of one

As we get wins in from a card, we increment the first `wins` of the
incoming queue of card counts by the number of the current card (since
we now know that is how many instances of the current card we have
played). Then we copy the entire queue back one to move onto the next
card