2023 Day 8
I tried to keep runtimes down by parsing the network map into a fixed
array of a pair of numbers in a tuple. The numbers represent the index
in the map to go to, depending on whether the instruction is a left or
right.

For part 2 I realised I needed to store the name of the nodes as well as
my minimal map. This was so I could find all my starts and also know
when we had got to a termination node (since I sorted part 1, I was only
ever going from the first element to the last element)

I was worried about how figure the end state, but it turned out that
Eric had been very kind to us and made each path only ever cycle from
just one end point, from that end point it only ever cycled back to the
same end point, the instructions were always on step one (which meant
the end point cycles were constant), and finally that the steps to get
from the initial start point to the first instance of the end point
were the same. That's a lot of luck, it's almost like Eric designed it
this way...

Anyway, having established all of that, part 2 reduced down to
calculating the LCM of each path to its first end point