2023 Day 9
Looking at this problem, I decided that there was likely to be a formula
to find the next element in a series like this. After a spot of pen &
paper I found the formula, which involved alternately adding and
subtracting the elements times by the combinations for picking that
many elements in out of the total number of elements.

I had to hard code the combinations as the real data had 21 elements on
each line and 21! (needed to work out the combinations factor)
overflowed a 64 bit number.

With the formula in hand this was quite straightforward. Part 2 only
required reversing the order of the elements