2023 Day 10
This was quite an involved day. I probably made things harder with all
the explicit type modelling, it certainly made things more verbose!

For part 1, I found the start point and then followed the pipe around,
counting as I went.

For part 2, I used the insight that one side of the pipe (measured from
the point of view of the travelled direction of the pipe) would always
be disjoint from the other, as there is no way to bridge the pipe since
it forms a continuous unbroken circuit. One of these sides will be the
inside and the other the outside. As we do our initial pipe pass for
part 1, we can take note of and store which side the adjoining tiles are
on. Not every non-pipe tile is adjacent to a pipe, so I filled these in
with a second pass as a flood fill from categorised side tiles. All that
remained was to work out which side was on the outside and count up how
many of the opposite side we had found.