2023 Day 12
Today was a dynamic programming day, which I understand to be recursion
with memoisation.

I initially did Part 1 with just recursion. This is because for any
pattern of springs and groups to be fitted, the number of matches could
be broken down into:

* Find the next match location for the next group to be fitted (or prove
  that it doesn't fit, in which case there are no matches for that input)
* If we do find a next match, the matches will be the sum of two smaller
  sub-problems:
  - the same set of groups, but the pattern truncated to omit the start
    up to and including the next match, so we force the algorithm to
    look for a subsequent match for the groups
  - the set of groups, with the first item removed looking on from
    immediately after the next match

This is a natural recursive pattern and worked well once I had
correctly accounted for the search termination cases (like skipping a
spring in the input pattern that is know to be damaged and so _must_ be
mapped to in a valid match)

This approach broke down for part 2, where it was obvious we were
searching the same patterns over and over again (created by the problem
statement of duplicating the input patterns to expand the question). So
out came a HashMap to memoise previously found instances of unique group
/ pattern combinations. Did the job nicely

I did try one big HashMap for every line of the input (the commited
solution uses one new HashMap per line). However it ran slower as there
was much less pattern duplication across lines than within each pattern
and the increased size of the one map was obviously the bigger
constraint